Created shift register circuitry and imported footprints
I think 5MHz to 10MHz is the theoretical maximum of this circuit. Actual
frequency will probably be around 5MHz or somewhat less given the signal
integrity issues on a 2-layer board with this many plane breaks.

Created footprints for button and HTSSOP

The HTSSOP footprint is special because I plan to hand solder, so I
added excessive thermal vias and a bare pad on the backside.

The netlist is imported to the PCB, but only the switches are placed.