Added device USB driver and HID implementation from simple-led-watch
Some modification was necessary because of the way the HSI48 is
implemented on the STM32F042, but mostly it is the same.

One oddity is that when the device is first programmed,  linux can't
find the interrupt endpoint. After a device reset or when the device is
plugged back in anew it can find the interrupt endpoint.