Added initial DMA-driven shift registers
There is a big problem: I made a mistake with LE and PL#. Apparently
when PL# is low, the shift register always outputs high. However,
whenever LE is high it will continually load the LEDs, so the user would
see weird LED patterns.

I think I can work around this by briefly turning off the LEDs. The
transactions take like 20us to complete.

Another issue is the fact that for some reason we are transferred 64
bits instead of 32.