Fixed bad wiring with a bad workaround
I drastically lowered how often we query the shift registers, but kept
them going as fast as possible. OE# is deasserted for the duration of
the transmission and PL#/LE is brought high so that the input register
actually gives us real data (side effect: the LED shift register is
outputting everything it sees on its serial input, hence why we
deasserted OE#). At the conclusion of the transmission, OE# is
re-asserted.