Fixed a rather subtle mistake with the midi queue
It wasn't properly asking if the queue was empty. It would only
partially flush the queue and only rarely actually fully flush it.

This caused buffers to be transmitted one after the other, which it
turns out actually bogs down the processor significantly. Stuff would
still respond, but very delayed.