Fixed the USB descriptor generator to handle the HID indexing
I did not realize that the HID report index needed to be the index of
the interface, but it makes sense. We had just so happened to always
compile in just the right order so the index was zero, but with the
bootloader addition that changed. For the bootloader, it is still
interface 0, but for the device it is interface 2. The descriptorgen
(which is already a horrible hack) was hacked further to support this.