Added initial bootloader python programmer script
I needed to add an explicit erase command since the erase size and the
program size are not the same on the STM32F0xx.

I added the initial implementation of the midi-fader resetting into the
bootloader, but I have not yet been successful in getting the actual
firmware to run. There is a bug right now with the script where it can't
handle dis-contiguous hex files for some reason.