Created configuration processor based on futures
The GUI thread (which is the main thread) is going to use a tokio mpsc
to send requests to this task which processes the requests as a stream.
Each request gets its own tokio response queue which will be manually
polled by the GUI thread until the result is complete.